make_factory_package: Create /etc/lsb-factory in diskimg if not exists
lsb-factory was created or copied into DUT when installed from
mini-Omaha server (or RMA shim). It was not required for diskimg mode.

However, recent factory pytests (start) may try to determine system
installation completeness according to the existence of lsb-factory.
As a result, we do want to create an empty lsb-factory in diskimg mode.

BUG=None
TEST=./factory_setup/make_factory_package.sh --diskimg diskimg.bin \
  --hwid_updater none --factory chromiumos_test_image.bin \
  --release recovery_image.bin

Change-Id: I0eedace5742a1389d4460af1ec766deae18f8500
Reviewed-on: https://chromium-review.googlesource.com/365478
Commit-Ready: Ting Shen <phoenixshen@chromium.org>
Tested-by: Ting Shen <phoenixshen@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>