installer: allow installing some subdirectories only
For station based testing, we don't need to install the entire factory
toolkit on DUT, but only some necessary files. For example, the 'init'
and 'sh' directory.
In this CL we add an option for installer to only install some of the
directories under '/usr/local/factory/' (and the 'enabled' file).

BUG=chromium:506788
TEST=run `./install_factory_toolkit.run chromiumos_test_image.bin \
          --install-dirs init sh --apps='-goofy'`
     verify that only factory/inin and factory/sh are installed and
     goofy is disabled

Change-Id: I7ac0c35464c52214c8426f668411f79f6956b416
Reviewed-on: https://chromium-review.googlesource.com/365486
Commit-Ready: Wei-Han Chen <stimim@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>