vswr: remove the USB feature.
Originally, we use the USB as a backup option to store of the config
file and log result when shopfloor connectivity is unstable. Now we
don't capture the screenshot of ENA, the storage requirement is not
huge. Therefore we choose to store them in local side.

BUG=b:30448439
TEST=manually

Change-Id: I0adb29442cb013184604361477c10c95623f0881
Reviewed-on: https://chromium-review.googlesource.com/366922
Commit-Ready: Chih-Yu Huang <akahuang@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Joel Kitching <kitching@chromium.org>