goofy: invocation: fix failure when resuming tests.
When Goofy is trying to resume tests (e.g., reboot test), it will fail
after the device is rebooted, because the dargs is not properly passed
to "invoke_pytest".

In CL:362740, we move the timing to resolve dargs before calling
invoke_pytest, and the resolved dargs should be passed to invoke_pytest.
However, when we are resuming the test, the variable that should be
passed to invoke_pytest is not initialized.

In this CL, read the resolved dargs from metadata and pass it to
invoke_pytest.

BUG=None
TEST=manual

Change-Id: I014dc85a61ae32d4a15599d10401cd9c81ff55d4
Reviewed-on: https://chromium-review.googlesource.com/367030
Commit-Ready: Wei-Han Chen <stimim@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Wei-Ning Huang <wnhuang@chromium.org>