file_utils: fix a bug in atomicwrite
When a relative path containing a subdirectory is provided to
AtomicWrite (for example: 'subdir/target'), the path is used
as the `prefix` argument of mkstemp.  However, when a relative
path is provided to mkstemp as the prefix, it automatically
prepends its temporary directory (on *nix-based systems,
usually '/tmp').  Thus, AtomicWrite ends up attempting to write
to the file '/tmp/subdir/target.TMP' and fails since the 'subdir'
directory does not exist.

The desired behaviour is to write to a temporary file in the
same directory as the target file.

Create a unittest to make sure this use case works properly.

BUG=None
TEST=Run file_utils_unittest.py

Change-Id: I53ebe796f8977a1737b4b97e211885b50b8f1d46
Reviewed-on: https://chromium-review.googlesource.com/369020
Commit-Ready: Joel Kitching <kitching@chromium.org>
Tested-by: Joel Kitching <kitching@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>