offline: add delay_after_reboot_secs args
Add an arguments to delay given seconds after each reboot before starting tests.
This is because on some devices the test need to wait some dependency services
to start. A better approach is to write individual rules for each test and
execute it before invocation, but for now we use a simple workaround to set a
fixed delay. This also helps to us to get currect timestemp, because it
also takes time for some devices to setup the device time after booting.

BUG=None
TEST=Manually

Change-Id: I2e0814e488ba097b17208330dddc21ce5305e1cb
Reviewed-on: https://chromium-review.googlesource.com/370979
Commit-Ready: Shun-Hsing Ou <shunhsingou@chromium.org>
Tested-by: Shun-Hsing Ou <shunhsingou@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>