type_utils: allow overriding property of parent class
When we are using LazyProperty.Override, if the property is defined in
parent class, then this property cannot be overriden.

  LazyProperty.Override(derived_class_instance,
                        'parent_class_prop',
                        stub_value)

Will raise 'has no attribute named "parent_class_prop"' error.

In this CL, we use 'hasattr' to check if a property is defined in a
class instead of using __dict__.  Also change to use 'getattr'.

BUG=None
TEST=python utils/test_utils_unittest.py

Change-Id: Ied0d624a53d6df6edf5a5733ac15af46314bbb78
Reviewed-on: https://chromium-review.googlesource.com/376579
Commit-Ready: Wei-Han Chen <stimim@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>