gooftool: add a waiver list to 'gooftool finalize' and 'gooftool verify'
For early builds, we may need to disable some functions in gooftool when
the underlying software/hardware is not ready. A workround to this is
commented out those functions in gooftool and maybe creates a 'DO NOT
COMMIT' CL like CL:363670.

To solve this problem, add a --waive_list that accept a list of items, and
for phase <= DVT, functions listed in the waive_list are skipped,
for phase >= PVT_DOGFOOD, this list should be empty.

Usage example:
  gooftool finalize --waive_list verify_tpm generate_stable_device_secret

Also make this option available in 'finalize' test.

BUG=None
TEST=Run finalize test.

Change-Id: I748a13ebcc04b20f00cf4969a42ee1668aa8c159
Reviewed-on: https://chromium-review.googlesource.com/377579
Commit-Ready: Ting Shen <phoenixshen@chromium.org>
Tested-by: Ting Shen <phoenixshen@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>