dome: front-end: allow only one task running at the same time
In order to provide the cancel function, we need to queue tasks. Tasks
are executed one-by-one, tasks on the waiting list are cancellable
(tasks that finished and running are not cancellable).

This patch doesn't implement the cancel function but implements the
queue mechanism (so we can implement cancel function in the future).

BUG=b:30999924
TEST=Manually tested

Change-Id: Ia7832f00194657ac62ffd9d34e9191e7c043ea3c
Reviewed-on: https://chromium-review.googlesource.com/384272
Commit-Ready: Mao Huang <littlecvr@chromium.org>
Tested-by: Mao Huang <littlecvr@chromium.org>
Reviewed-by: Wei-Han Chen <stimim@chromium.org>