dome: set maximum process to 1 in uwsgi.ini
This is the lazy way to make sure no racing condition will happen. Say,
if we made a change to bundle A, Dome stages the Umpire config, tries to
update, and deploy. If we tried to make a change to bundle B before task
A finishes, task B won't be able to stage the config since it's already
staged, thus, task B will definitely fail. We can implement a
synchronization mechanism in the back-end, but that would take some
effort, let's just make it 1 unless we really need to increase
performance.

BUG=b:30999924
TEST=Manually tested

Change-Id: I7b5914fefdfd089c7d831c8f52ffdbff13061e90
Reviewed-on: https://chromium-review.googlesource.com/384273
Commit-Ready: Mao Huang <littlecvr@chromium.org>
Tested-by: Mao Huang <littlecvr@chromium.org>
Reviewed-by: Youcheng Syu <youcheng@google.com>