dome: fix a bug that Umpire does not actually get files uploaded by Dome
Uploaded files are stored in /tmp by default, which is not shared
between Dome and Umpire Docker containers currently. Therefore, Umpire
cannot see uploaded files from Dome.

A coincidence is that we forgot to remove the factory toolkit file used
when setting up a new Umpire container. The file was left in the
container. Every time we updated factory toolkit, the Umpire container
read that old file instead of the new file we uploaded, creating an
illusion that the update succeeded while it's not.

This patch also works around b/31415816 in which the temporary files
were not removed immediately after finished requests.

BUG=b:31417203,b:31415816,b:30999924
TEST=Manually tested

Change-Id: I016178f24c87b17ef27390b21787410c55191c12
Reviewed-on: https://chromium-review.googlesource.com/384095
Commit-Ready: Mao Huang <littlecvr@chromium.org>
Tested-by: Mao Huang <littlecvr@chromium.org>
Reviewed-by: Wei-Han Chen <stimim@chromium.org>