goofy: connection_manager: fix connect to WLAN
When goofy is trying to communicate with shill using dbus-python, and
introspect gives no result about method signature, it would guess the
type to use if no signature is given. For dictionary type, it would
use an arbitrary entry in it, and use its type as the type of all
entries.

Since the org.chromium.flimflam.Manager interface seems to be not
introspectable, and we have both boolean and string as dictionary values,
we need to explicitly specify the signature.

BUG=chromium:645870
TEST=manual

Change-Id: I123fbd0658b8d4b889eabc093e2e32c25c7c998a
Reviewed-on: https://chromium-review.googlesource.com/384079
Commit-Ready: Pi-hsun Shih <pihsun@google.com>
Tested-by: Pi-hsun Shih <pihsun@google.com>
Reviewed-by: Shun-Hsing Ou <shunhsingou@chromium.org>