dome: front-end: add boards app (originally the WelcomePage)
The purpose of WelcomePage was letting the user to select board. Once
selected, the user wasn't able to go back to the WelcomePage to select
other boards previously. This is a problem since the user will need to
change board some times.

Making this page a new app solves the problem, also unified the
behavior of all apps (WelcomePage was a special "board selection app"
before).

This patch:
- Renames WelcomePage to BoardsApp
- Merges AppPage into DomeApp

BUG=b:30999924
TEST=Manually tested

Change-Id: Icbbe3b9f4a4cd5cced2d38246bbcfdf5e8d8bae1
Reviewed-on: https://chromium-review.googlesource.com/386544
Commit-Ready: Mao Huang <littlecvr@chromium.org>
Tested-by: Mao Huang <littlecvr@chromium.org>
Reviewed-by: Wei-Han Chen <stimim@chromium.org>