Implement the yaml representer of the region field and component.
In this CL, we can dump the RegionField and RegionComponent to
yaml. Note that we don't support the legacy style of RegionField.

Because of the restriction of PyYaml, there is a dummy string
after the RegionComponent. But it does not affect the result of
yaml loading. That is, when we load a RegionComponent, dump it
back to yaml string, and then load it again. The Python instance
is the same.

BUG=chromium:649246
TEST=py/hwid/v3/yaml_tags_unittest.py

Change-Id: I127dbba632b33f012453e6fd028c37fe83e5ce8e
Reviewed-on: https://chromium-review.googlesource.com/388371
Commit-Ready: Chih-Yu Huang <akahuang@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>