yaml_utils: Parse mapping to OrderedDict.
While we load a yaml file to a object, modify the object, and dump to
a yaml file, we hope to keep the order of the mapping instance.
Therefore, we should parse the mapping to the Python OrderedDict
object, and dump the OrderedDict instance to yaml like a dict object.

BUG=chromium:649246
TEST=py/utils/yaml_utils_unittest.py

Change-Id: Id5da27765bdc9a9b804b1171e268c2983c7170b4
Reviewed-on: https://chromium-review.googlesource.com/388250
Commit-Ready: Chih-Yu Huang <akahuang@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>