process_utils: add ppid to IsProcessAlive
When we checked if a process is alive or not, only pid was given.
However, if the process is actually terminated, and unfortunately a new
process is running with the same PID, then we are checking on an
irrelevant process.

By providing ppid (parent PID) of given process, we can be more
confident that we are watching the correct process.

The ppid argument is optional, without this argument, the function
checks process status only.

BUG=None
TEST=manual

Change-Id: I0da47b65fba899b4a73653803bc26085e1474a98
Reviewed-on: https://chromium-review.googlesource.com/388411
Commit-Ready: Wei-Han Chen <stimim@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>