dome: front-end: allow task to start only if all tasks before succeeded
There are dependencies between tasks: a task depends on all tasks before
it. If a task failed, all following tasks may fail. For example, if the
user tries to upload a new bundle, and update its factory toolkit. It's
impossible to run the 2nd task if the 1st failed since the bundle won't
be there at all. Therefore, we should not allow a task to run if any of
the tasks before it has failed.

BUG=None
TEST=Manually tested

Change-Id: Ie796fcfe45bd16f8feac668c79146aff3e289f73
Reviewed-on: https://chromium-review.googlesource.com/390592
Commit-Ready: Mao Huang <littlecvr@chromium.org>
Tested-by: Mao Huang <littlecvr@chromium.org>
Reviewed-by: Youcheng Syu <youcheng@google.com>