dome: back-end: add POST /files/ API
The API of Dome is designed as: if a request contains N files, split it
into (N+1) requests, with first N request uploading each file (encoded
in multipart/form-data), and the last request sending other fields
(encoded in JSON). Each file uploading request gets a file ID, in the
last request, we refer to a file using its ID we got before.

For example, if we want to send a request:
{
  "foo": "bar",
  "hello": "world",
  "file01": a file object,
  "file02": another file object
}
We split it into 3 requests:
1. Send {"file": a file object} in multipart/form-data, suppose we got
   ID 1005.
2. Send {"file": another file object} in multipart/form-data, suppose we
   got ID 1008.
3. Send
   {
     "foo": "bar",
     "hello": "world",
     "file01": 1005,
     "file02": 1008
   }
   in JSON.

This CL is preparation for the change.

BUG=b:30999924
TEST=Manually tested

Change-Id: I64c57ea40d6b6a947c9e4e7e2a134e815960bbfa
Reviewed-on: https://chromium-review.googlesource.com/391531
Commit-Ready: Mao Huang <littlecvr@chromium.org>
Tested-by: Mao Huang <littlecvr@chromium.org>
Reviewed-by: Ting Shen <phoenixshen@chromium.org>