Makefile: Change overlay search logic to use Portage command.
Searching overlay with hard-coded logic is causing problems when we need
to deal with overlay design changes (like the convention of using
project-*) or difference between private and public overlays.

The right way to find package is using equery-$BOARD which package, that
will provide the complete path of ebuild files and can be used to find
files folder.

BUG=chromium:577601
TEST=make overlay-gale

Change-Id: Iac1024f4b508b306dba694759f67f677e9585a1b
Reviewed-on: https://chromium-review.googlesource.com/392769
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Wei-Han Chen <stimim@chromium.org>