dome: adopt djangorestframework-camel-case
The naming convention is lowercase_separated_by_underscores in python,
but CamelCased in JavaScript. This creates issues between the
communication of front-end and back-end. One of them needs to adopt the
other, resulting in inconsistent naming style in code (currently it's
the front-end adopting the back-end).

This package automatically converts JSON requests from CamelCased to
lowercase_separated_by_underscores and vice versa, eliminating the
problem once and for all.

BUG=b:30999924
TEST=Manually tested

Change-Id: Iab86af75e6069328cb1ac2ab0bcffb1e9157ff1f
Reviewed-on: https://chromium-review.googlesource.com/390881
Commit-Ready: Mao Huang <littlecvr@chromium.org>
Tested-by: Mao Huang <littlecvr@chromium.org>
Reviewed-by: Ting Shen <phoenixshen@chromium.org>