hwid: remove binary_string and encoded_string in HWID class.
Originally, HWID has the database, BOM, binary string, and encoded
string. But the binary string actually depends on the database and
the BOM only, and the encoded string is just another representation
of the binary string. So now we only keep the database and the BOM
object in HWID, and calculate the binary string while needed. We can
also ignore the verification between BOM and the binary string.

If the HWID is created from the encoded string, we still store the
original one, and return it if the BOM is not changed.

BUG=chromium:653369
TEST=unittest in py/hwid/v3

Change-Id: Ic9f7d203c07ebb481c412303ce273fa5c4afbb07
Reviewed-on: https://chromium-review.googlesource.com/396638
Commit-Ready: Chih-Yu Huang <akahuang@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>