dome: add Makefile for linting
For the back-end (python), we can use the same way to lint as other
files in the factory repo would normally do. We'll simply delegate the
linting task to factory repo's Makefile.

But for the front-end, we can't take the same approach since we're using
React.

BUG=None
TEST=Manually tested

Change-Id: I330e2d84751915bd0336eb069e4bde7ee932d1d9
Reviewed-on: https://chromium-review.googlesource.com/394387
Commit-Ready: Mao Huang <littlecvr@chromium.org>
Tested-by: Mao Huang <littlecvr@chromium.org>
Reviewed-by: Youcheng Syu <youcheng@google.com>