Makefile: Introduce 'BOARD'* variable and add 'overlay' target.
Currently there are many pseudo targets (overlay-*, ...) that assumes
the postfix of target is the name of board; this can be changed by
introducing a real BOARD variable so we can determine the BOARD and
related configuration (or files) easier - for example finding the
chromeos-factory-board/files directory from private overlays.

This change adds BOARD and BOARD_FILES_DIR and adds a generic 'overlay'
target so overlay-* can use it.

BUG=chromium:577601
TEST=make overlay-gale; make overlay BOARD=gale

Change-Id: I0d46ca68acb91868bbf2e29b1ab5f19f9c1d47bd
Reviewed-on: https://chromium-review.googlesource.com/394568
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>