Makefile: Change 'par' target to use new resource system.
The 'par' used to rely on build options to decide its contents and
output name. However we've recently realized that
 - Usually it's fine to put complete par (4M) instead of factory-mini.par (1M).
   In fact, the difference between factory-mini.par and factory.par is
   actually causing problems.
 - factory.par is widely used in station-based testing.

As a result, a long term goal is we'll eliminate factory-mini.par to
simplify build process, especially that we have the 'resource system'
now.

The first step is to migrate 'par' target to build with resources, and
because of that, we can now always build both factory.par and
factory-mini.par, no need checking MAKE_ARGS or PAR_NAME.

BUG=chromium:577601
TEST=make par; make par BOARD=gale

Change-Id: I5cf6342717ce18a16b875b4bf823984749dc1350
Reviewed-on: https://chromium-review.googlesource.com/394395
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>