goofy: change run_if behavior
For a FactoryTest, if run_if argument is set to 'foo.bar', then Goofy
will skip the test if 'foo.bar' is set to False in shopfloor aux data.

If 'foo.bar' is not set, then Goofy would *not* skip the test.  However,
in current usage, this behavior is annoying.  For example:

  FactoryTest(
      ...
      run_if='device_data.component.HAS_LTE',
      ...)

This test will not be run if and only if 'HAS_LTE' is set, and is set to
False.  And in most of projects, they want the test to be skipped if the
value is not set.

In this CL, if key is not found in aux data, the query result is
considered None (or False).  Therefore, in the same situation above,
when HAS_LTE is not set, the test will be skipped.

Another use case for run_if is:

  FactoryTest(
      ...
      run_if='!device_data.complete_smt',
      ...)

If complete_smt is not set, the test *will* still be run in new
implementation.

BUG=b:31678190
TEST=make test

Change-Id: I5a9c9ee446bdd004e532e2bf3f8b2ffa2b7b46b3
Reviewed-on: https://chromium-review.googlesource.com/394511
Commit-Ready: Wei-Han Chen <stimim@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>
Reviewed-by: Shun-Hsing Ou <shunhsingou@chromium.org>