lumberjack: disable archiver_unittest
The test failed because of gpg runtime error when doing decrypt, which
is because of some gcc optimization bug that cause the built gpg to
malfunction. (CL:394427)
Since we're going to deprecate lumberjack soon, and we're going to
migrate to clang+llvm, this should not be a big issue. Disable the
unittest so "make test" can pass.

BUG=None
TEST=make test

Change-Id: I01ccd89f47e9713de026dd0e4d82d1560e10b581
Reviewed-on: https://chromium-review.googlesource.com/394449
Commit-Ready: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>