hwid: change logic of 'Database.VerifyBOM' method.
- Only verify the components in the corresponding image_id
Because the components for each image ID might be different, we only
verify the components in the target image ID, not all components
listed in the database.

- Add 'probeable_only' argument
When the BOM is decoded by HWID string, it would contain every
component in the database. But if the BOM object is created by the
probed result, it does not contain the unprobeable component before
evaluating the rule. We should verify the probeable components only.

BUG=chromium:653369
TEST=database_unittest.py

Change-Id: Iceb60c2b72a64e7045d5013aeea5248235e31e3e
Reviewed-on: https://chromium-review.googlesource.com/397201
Commit-Ready: Chih-Yu Huang <akahuang@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Wei-Han Chen <stimim@chromium.org>