hwid: Change the logic of transfer probe result to BOM.
For the unprobeable components class, usually we would set the value
while evaluating the rules. But if there is only one item in the
component class, there might be no rule to set the value because
originally the index of the encoded field is set to 0 as default.

Now we don't set the default index to 0. To handle this special
situation, if an unprobeable component class only has one item, we
pretend we found this item.

BUG=chromium:653369
TEST=unittest in py/hwid/v3

Change-Id: I0bbc2e1787a44c163851797ed0872bbcc1d3f5c1
Reviewed-on: https://chromium-review.googlesource.com/397200
Commit-Ready: Chih-Yu Huang <akahuang@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>