hwid: BOM only keeps the known components.
Orginally when we decode the HWID string, we set the index of missing
encoded fields to 0. This is because we needed to encode BOM before we
evaluate the unprobeable components.

Since now we can encode the BOM at the last minute, i.e. after it gets
the information of image_id and the unprobeable components, we dont
set the index of the unknown field as 0.

BUG=chromium:653369
TEST=unittest in py/hwid/v3

Change-Id: I8e19c2917d7b3502b3bbe1c03dac62614ef337d6
Reviewed-on: https://chromium-review.googlesource.com/397202
Commit-Ready: Chih-Yu Huang <akahuang@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Wei-Han Chen <stimim@chromium.org>