Makefile: Allow building closure out-of-tree.
In portage system, if build intermediate won't need to live in source
tree then portage won't need to copy source to a temporary folder first
and can simply build with original copy of source. This is called
"out-of-tree build".

To support building in out-of-tree, we want to migrate the only item that
writes output in source folder -- closure. If the new variable
OUTOFTREE_BUILD is set to non-empty value then we will ask closure to
output in build/closure.

BUG=chromium:577601
TEST=make

Change-Id: I5dc7cbb95d0651b05a907a893b6f54d39e85589a
Reviewed-on: https://chromium-review.googlesource.com/397085
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Shun-Hsing Ou <shunhsingou@chromium.org>