Makefile: Generalize extraction of external dependencies.
Some resource files may live in external website and must be prepared
either in build process or before 'make' starts (for example, portage
may fetch and extract webgl_aquarium_static in a different folder).

To allow better integration and to keep development easier, we should
generalize the preparation of external dependency. The new
func-extract-from-url can be used to download and extract a given gs://
URL. Targets calling this function should define a variable as folder to
extracted contents (so if there's build system like portage it can
pre-download and extract to that folder).

BUG=chromium:577601
TEST=make closure

Change-Id: I2266f3f79e02d38ed36ac5d1f5ff51270d1b7dd1
Reviewed-on: https://chromium-review.googlesource.com/397086
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Shun-Hsing Ou <shunhsingou@chromium.org>