hwid: add "default" argument in component field.
we introduce a new flag: 'default' in the component field. When the
component is 'default', it will be always matched. Moreover, if
another component is also matched, this default component will be
ignored.

The flag can solve the problem that we need to set "!re .*" at early
stage and modify to a fixed value later.

BUG=chromium:655924
TEST=py/hwid/v3/database_unittest.py

Change-Id: I7e9e23906c48ae3523683cbd37fc43ec1274c953
Reviewed-on: https://chromium-review.googlesource.com/398686
Commit-Ready: Chih-Yu Huang <akahuang@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Wei-Han Chen <stimim@chromium.org>