config_utils: follow symbolic link when using default search path.
When running config_utils.LoadConfig without specifying default search
path, we use the caller module path by default. However, if the module
is a symbolic link, we may also want to read config from the
directory of its original file.

This CL adds this behavior. We search the config in the original
directory first, and override it by the config beside the symbolic link
if exists.

BUG=None
TEST=1. config_utils_unittest.py
     2. Put config in py/goofy, and read config in goofy.py by executing
     bin/goofy.

Change-Id: I884bb8b771aa9e12ce2a5e3051f4911959bb7e3d
Reviewed-on: https://chromium-review.googlesource.com/401100
Commit-Ready: Shun-Hsing Ou <shunhsingou@chromium.org>
Tested-by: Shun-Hsing Ou <shunhsingou@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>