goofy: initial implementation of backend plugin
Currently goofy has a lot of platform dependent components. Some of the
components are written as functions in goofy, while others are separated
classes. To make it easier for user to decide what to use, we want to
create an unified interface for all platform dependent parts, that is,
goofy plugins. The core of goofy should only do test flow control.

This CL creates the basic implmentation of backend goofy plugin.
Moving of code in goofy to plugins will be uploaded in following CLs.

BUG=chromium:658151
TEST=make test

Change-Id: Ib17e8fd62f14c5a5135c0bd5da91345cdb0e3118
Reviewed-on: https://chromium-review.googlesource.com/400322
Commit-Ready: Shun-Hsing Ou <shunhsingou@chromium.org>
Tested-by: Shun-Hsing Ou <shunhsingou@chromium.org>
Reviewed-by: Wei-Han Chen <stimim@chromium.org>