Makefile: Prevent checking build_packages
It is not recommended to detect if we're building under build_packages since it
was supposed to be a simple wrapper for emerge commands.

This change tries to solve the problem differently: consider "if the build
system instructs to not pause = not an interactive bug-fix cycle, then don't
check board resource timestamps".

Developers calling 'make' manually will still perform the right checks.

BUG=chromium:659874
TEST=touch chromeos-factory-board-*.ebuild
     emerge-$BOARD chromeos-factory
     ./build_packages --board $BOARD

Change-Id: Id81d5ea90fef4b7f97cf17db47338be15b61a536
Reviewed-on: https://chromium-review.googlesource.com/407129
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Shun-Hsing Ou <shunhsingou@chromium.org>