goofy: TestListIterator: handle run_if
The FactoryTest object can have a 'run_if' attribute, which can be used
to turn on / off a test.  For example::

  FactoryTest(id='a', run_if='foo.a')

Then test 'a' will run if key 'a' is found in table 'foo' in shopfloor
aux data, and the corresponding value is equal to True.

If test 'a' runs before test 'b', and test 'b' has run_if='bar.b', then,
we can set the value of 'bar.b' in test 'a' to change the test list
behavior.  Since the run_if function is evaluated right before a test
gets its chance to run, we don't need to call "UpdateSkippedTests" to
update the list of skipped tests when we modify the aux data.

To make things simpler, a test cannot prevent its siblings from running
by making the run_if value of its parent to False.  For example::

  with FactoryTest(id='G', run_if='foo.G'):
    FactoryTest(id='a')
    FactoryTest(id='b')

Changing the value of 'foo.G' in test 'G.a' won't stop test 'G.b' from
running.

BUG=b:32721069
TEST=python goofy/goofy_unittest.py TestListIteratorTestSuite

Change-Id: I60fe7e8f046e0a08b511a018eaca14132deca18e
Reviewed-on: https://chromium-review.googlesource.com/409852
Commit-Ready: Wei-Han Chen <stimim@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>
Reviewed-by: Wei-Han Chen <stimim@chromium.org>