umpire: make ip in config optional.
Since we're using umpire in docker now, we can select which interface to
bind (if needed) on docker level, and we can set binding ip to 0.0.0.0
inside docker. This also reduce the chance of accidentally set a wrong
ip in config file.

CQ-DEPEND=CL:408502
BUG=None
TEST=make test
TEST=manually tested, umpire_docker.sh build, umpire_docker.sh start,
     open shell into docker and check umpire server is working.

Change-Id: I9f6a9b530dbfca97c9324e7f6f40456f4209d730
Reviewed-on: https://chromium-review.googlesource.com/408544
Commit-Ready: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>