umpire: change umpired to be started by docker directly
umpired was started by upstart before, which cause some hack to make it
work with docker. Since now umpire is already in the docker image, we
can run it as docker CMD directly.
There's another service (tftpd) that is started by upstart, but since
it'll need to be one service for all different instance of umpire, we'll
need to pull it out. tftp does not work now.

CQ-DEPEND=CL:403577
BUG=None
TEST=make test
TEST=manually tested, umpire_docker.sh build, umpire_docker.sh start,
     open shell into docker and check umpire server is working.

Change-Id: I606b251662f29753f1c73523df10c1a70a0401f1
Reviewed-on: https://chromium-review.googlesource.com/403717
Commit-Ready: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Mao Huang <littlecvr@chromium.org>