umpire: remove fastcgi_server.WSGISession
There are two WSGISession in umpire: in web.wsgi and in fastcgi_server.
Remove the one in fastcgi_server and change usage to the one in
web.wsgi.

There's actually lots of duplicate codes in fastcgi_server and in
web.wsgi, web.xmlrpc, daemon. The plan is to remove most things in
fastcgi_server and merge it with other codes.

BUG=None
TEST=make test
TEST=manually tested, apply the patch to active device toolkit inside
     umpire, restart umpire. Check RPC call to umpire for shop_floor,
     rpc_dut and rpc_cli are working. Check /resourcemap webapp is
     working.

Change-Id: I316e0335216de4301ad9f7b71c55c815b9e2e44e
Reviewed-on: https://chromium-review.googlesource.com/409872
Commit-Ready: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>