umpire: fix daemon_unittest.py
The test were always using default port 8081, which cause it to fail
when the port is occupied. Change it to use unused port.

Singleton for daemon is not useful now and would make testing harder
since it force different tests to use the same instance of daemon.

Also, the testWebAppSite test never test if the result returned is
actually expected, and there's a bug in TestWebApplication. Fix this
by actually checking if the returned http body is same as expected,
and don't set callback for error case so that error case would make
test fail.

BUG=None
TEST=make test
TEST=run LOG_LEVEL=1 py/umpire/daemon_unittest.py, check the debug
     output is as expected.

Change-Id: I07fbc20921d9c5bc945419e8f50a93bdf22dfda1
Reviewed-on: https://chromium-review.googlesource.com/411080
Commit-Ready: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>