deployment: add "publish", "save" commands, and switch to xz
The "save" command saves the Docker images to the current working
directory.

The "publish" command builds, saves, and uploads the Docker image to
chromeos-localmirror. We'll have to manually publish the Docker image
before the automatic way is implemented. It is a tedious work, so adding
this helper function should be beneficial.

Also, switch to using xz instead of bzip2 since it provides better
compression rate.

BUG=b/32229544
TEST=./dome.sh save && ./umpire_docker.sh save &&
    ./dome.sh publish && ./umpire_docker.sh publish

Change-Id: Iafd257301d932612cae3b83698fe41bb576fb80e
Reviewed-on: https://chromium-review.googlesource.com/411549
Commit-Ready: Mao Huang <littlecvr@chromium.org>
Tested-by: Mao Huang <littlecvr@chromium.org>
Reviewed-by: Youcheng Syu <youcheng@google.com>