umpire: fix stopping lighttpd kills everything
We need to restart lighttpd when http config changes, which is
accomplished by stopping and starting the service.

In the original code, since twisted is using fork+execve to execute
lighttpd, and lighttpd is running in non-daemon mode, lighttpd would be
in the same process group with umpired. But when we send SIGTERM to
lighttpd, since lighttpd itself is a forking server, it would have many
children that it needs to kill, so what it do is call kill(0, SIGTERM)
when it is terminated by SIGTERM, which kills EVERYTHING in the same
process group, including umpired and all other services.

To fix this, we should call os.setpgrp() after we fork and before execve
lighttpd, but since that code is hidden deeply inside twisted, and
there's no hook that can modify the behaviour, we can't do this easily.

Another possibility is to do os.setpgid(child_pid, child_pid) after
reactor.spawnProcess, but this cause race condition since this command
would need to be executed before execve to work.

The workaround is actually calling reactor.spawnProcess with
usePTY=True, which cause twisted to use PTYProcess, which calls setsid()
before execve, and start a new session (thus a new process group) for
the child service.

This triggers another bug since the handling for process ending for
PTYProcess and Process is different, and current code stop twisted from
installing a proper SIGCHILD handler. Also fix that by let twisted
install the handlers, and install custom handler for SIGTERM / SIGINT in
twisted OnStart callback.

BUG=None
TEST=make test
TEST=manually tested, apply the patch to active device toolkit inside
     umpire, restart umpire. Check RPC call to umpire for shop_floor,
     rpc_dut and rpc_cli are working. Check /resourcemap webapp is
     working. Deploy something that would cause http to restart (adding
     service absub_logparser would do), and check deploying is
     successful.

Change-Id: Ia64f5dca8ab08eae23636a704ca016ab5263575e
Reviewed-on: https://chromium-review.googlesource.com/411649
Commit-Ready: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>