probe: Implement function class.
A instance of the Function class is a callable object. The input and
output are both list of dictionaries. We implemented the basic
structure in this CL. The new function should be implemented at
`probe/functions/<function_name>.py`.
After calling `function.LoadFunctions`, all the functions in the
`probe/functions/` folder are registered.

BUG=chromium:667651
TEST=unittest in py/probe/function

Change-Id: I2a4e814068055a01216d2a752df16c6df82649b5
Reviewed-on: https://chromium-review.googlesource.com/413166
Commit-Ready: Chih-Yu Huang <akahuang@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Wei-Han Chen <stimim@chromium.org>