probe: Implement probe command line.
We implemented two commands in probe command line:
- "eval_function"
Evaluate a probe function directly. It is handy for testing the
function implementation and finding the probed results. For example,
to call the "file" function to probe /tmp/test file:
$ ./probe_cmdline.py eval_function file /tmp/test

- "probe"
Probe the result according to the probe statements. It is similar to
"gooftool probe". The command requires a config file and output the
matched probe results. For example:
$ ./probe_cmdline.py probe probe_config.json

BUG=chromium:667651
TEST=py/probe/probe_cmdline_unittest.py

Change-Id: Ic1f693d3e3739f4f4eafc5bcb4b93bb998b26a33
Reviewed-on: https://chromium-review.googlesource.com/414669
Commit-Ready: Chih-Yu Huang <akahuang@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Joel Kitching <kitching@chromium.org>