init: add delayed installation script and document
During the reimaging stage, the DUT downloads a test image and a factory
toolkit to the local storage. The DUT is responsible for installing the
toolkit itself afterward. But in that stage, the DUT (running factory
shim or factory netboot installer) lacks of `python` to finish essential
works for installing the factory toolkit. It can only extract the
toolkit file but not install. We have to delay this work until the first
boot using test image environment.

BUG=b:33094336
TEST=Manually tested:
    1. Flash a DUT to test image (no factory toolkit installed)
    2. Copy the `init` folder to `/usr/local/factory/init` on the DUT
    3. Copy factory toolkit to `/usr/local/factory` on the DUT
    4. `echo "/usr/local/factory/install_factory_toolkit.run"
         >/usr/local/factory/init/install`
    5. Reboot the DUT, saw it installs the factory toolkit and invokes
       factory software

Change-Id: I89e0d7ca037f3905115fb5e76c6046a44691e4a4
Reviewed-on: https://chromium-review.googlesource.com/417640
Commit-Ready: Mao Huang <littlecvr@chromium.org>
Tested-by: Mao Huang <littlecvr@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>